Ability to detect harvest extension and read time since harvest
It took a while, but I was able to find a way to detect if Base Harvest or Biomass Harvest was initialized. For this, I just take a look at a variable that has to be initialized in those extensions (Landis.Library.HarvestManagement.SiteVars.TimeOfLastEvent). If it's null, then the extension hasn't been loaded. It's the same variable I use to see how long it has been since the site has been harvested, to know if a road should be constructed.