Trying individual road splitting to imitate flux algorithm from QGIS
I've tried to use an approach of splitting the pixelated roads into individual segments in order to imitate the flux algorithm from QGIS. However, it's not working very well. There are so many cases of pixel configuration that it is very hard to write an algorith, that can clearly detect how to individualize them. I've given up on this idea, and went for a simpler one.