Update to adapt to the new version of the harvest libraries
A problem that I think was due to the new harvest management libraries was that I couldn't read the "TimeSinceLastEvent" variable for the sites; it's like it wasn't initialized anymore. Thus, I switched to a new way of calling this variable, by initializing a site variable object and using the "GetSiteVar" function of the modelcore. It works like a charm.