Taking into account harvest repetitions
Before, the module didn't take into account the fact that when building a road to a recently harvested pixel, this very pixel might be harvested again soon by a repeated prescription(single or multiple repeat).

Now, the module will detect if the pixel will be harvested again by a repeated prescription, and will check when the next harvest will happen. It will then try to find what is the best option : to construct a costlier road that will last until the next harvest ? Or to create a small road that will be destroyed before the next harvest, and to re-create it when the time comes ? To do that, the module simply checks if the multiplicative cost of construction is lower than 2 for the road types that last until the next rotation. It will then choose the cheapest option.