Solving the bug with flux paths
There was a bug with the wood fluxes that resulted in single pixels or paths that fluxed twice or more, making for some weird patterns in the road types. It as due to the fact that I forgot to take out the cell of connection to the next flux path while giving the list of cells in a new flux path. It's now resolved.