Temporary fix in GetTimeBeforeNextHarvest()
Recent harvest extension end up giving a null object when trying to retrive the stand object associated with a cell, leading to the error described here : https://github.com/LANDIS-II-Foundation/Library-Harvest-Mgmt/issues/28. This commit makes a temporary fix by not taking into account the return to the cell if the stand object cannot be retrived.