Adapt to different writing of "none" for optional rasters
In the situation when "None" was indicated as the path for a raster file, the code didn't try to read the raster map, but still took into account the cost of the feature when computing the cost map. This was because of case-sensitive comparison to the word "none" in the code. Now, the comparison is case insensitive.